Add PlayerGameFinder class with StatFinder parent and GamelogTranslator
The StatFinder superclass includes HTTParty, sets the base_uri for queries,
and provides a simple alias for `get`, which will be useful for any type
of future stat queries that may arise (i.e. team stats, player season
stats, etc).

PlayerGameFinder hits the `/playergamelog` endpoint and queries for all
of a given player's games. Because the API only returns regular season
or playoff games (but not both) for a given query, PGF sticks an extra
key-value pair in the parsed gamelog hash to flag it as one or the
other.

GamelogTranslator encapsulates the translation of nba.com's stats to
match the attributes of `Game`, so that the model doesn't need to know
about the implementation of a foreign API.