Add Player#fetch_unstored_games and Game#build_from_gamelog methods
A Player instance uses a Stats::PlayerGameFinder to grab all
gamelogs for that player. For each that doesn't already exist in the DB,
it has Game build a new instance from that gamelog and saves it.

Game builds new instances using a Stats::GamelogTranslator so it doesn't
need to know about the differences between the API's implementation and
its own.